RF: Centralize submodule addition logic in GitRepo
Essential verification check was only available in `add`
